Fixed requests marco
